combined traits6 & traits7 into traits6 with compile
