+++ exercises/traits/traits6.rs
@@ -6,11 +6,17 @@
 
 // I AM NOT DONE
 
+use std::fmt::{self, Debug, Formatter};
+
+/// Struct representing a house
+#[derive(Default)]
 struct House {
     area_sqft: u32,
     purchase_date: String,
 }
 
+/// Struct representing a vehicle
+#[derive(Default)]
 struct Vehicle {
     name: String,
     model: String,
@@ -54,58 +60,17 @@ fn foo(flag: bool) -> ?? {
     }
 }
 
-#[cfg(test)]
-mod test {
-
-    use super::*;
-
-    fn init() -> (House, Vehicle) {
-        let house = House {
-            area_sqft: 5000,
-            purchase_date: "21 Nov 2017".to_string(),
-        };
-        let vehicle = Vehicle {
-            name: "BMW".to_string(),
-            model: "320d".to_string(),
-            purchase_date: "13 Aug 2022".to_string(),
-        };
-
-        (house, vehicle)
+impl Debug for dyn Details {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+    // print the summary of the struct returned from the function `foo`
+        write!(f, "{}", ??) // TODO: Complete the code
     }
+}
 
-    #[test]
-    fn check_foo_returns_house_if_true() {
-        let (house, _) = init();
-        assert_eq!(house.summary(), foo(true).summary());
-    }
-
-    #[test]
-    fn check_foo_returns_vehicle_if_false() {
-        let (_, vehicle) = init();
-        assert_eq!(vehicle.summary(), foo(false).summary());
-    }
-
-    #[test]
-    fn check_purchase_date_for_house() {
-        let (house, _) = init();
-        assert_eq!(
-            format!(
-                "The house has an area of {} sqft and was purchased on {}",
-                house.area_sqft, house.purchase_date
-            ),
-            house.summary()
-        );
-    }
-
-    #[test]
-    fn check_purchase_date_for_vehicle() {
-        let (_, vehicle) = init();
-        assert_eq!(
-            format!(
-                "The {} vehicle with model {} was purchased on {}",
-                vehicle.name, vehicle.model, vehicle.purchase_date
-            ),
-            vehicle.summary()
-        );
-    }
+pub fn main() {
+    let x = foo(true);
+    println!("{:?}", x);
+    // TODO: Complete the code
+    // print the summary of the struct returned from the function `foo`
+    println!("{}", ??);
 }
