+++ info.toml
@@ -740,23 +740,15 @@ See the documentation at: https://doc.rust-lang.org/book/ch10-02-traits.html#spe
 [[exercises]]
 name = "traits6"
 path = "exercises/traits/traits6.rs"
-mode = "test"
+mode = "compile"
 hint = """
 To ensure a function is able to return incompatible types can be done by using a trait bound that is common to them.
 Wrap the shared trait (implemented by the 2 struct types/objects) as a Boxed trait object. Try replacing the '??' with Box<_> 
-
 See the documentation at: https://doc.rust-lang.org/book/ch10-02-traits.html#returning-types-that-implement-traits
-"""
 
-[[exercises]]
-name = "traits7"
-path = "exercises/traits/traits7.rs"
-mode = "compile"
-hint = """
-To make the function printable on console using `:?` symbol. Use the `Debug` trait for the `Details` trait.
+Also, in order to make the function printable on console using `:?` symbol. Use the `Debug` trait for the `Details` trait.
 Here, the trait object has to implement `Debug` trait, otherwise the function returning trait object will 
 not be printable on console using {:?} symbol.
-
 See the documentation at: https://doc.rust-lang.org/std/fmt/trait.Debug.html
 """
